+++ src/loader.rs
@@ -2,28 +2,52 @@
 
 use std::collections::BTreeMap;
 
-use saphyr_parser::{Event, MarkedEventReceiver, Marker, Parser, ScanError, TScalarStyle, Tag};
+use hashlink::LinkedHashMap;
+use saphyr_parser::{Event, MarkedEventReceiver, Marker, ScanError, TScalarStyle, Tag};
 
 use crate::{Hash, Yaml};
 
-/// Main structure for quickly parsing YAML.
+/// Main structure for parsing YAML.
 ///
-/// See [`YamlLoader::load_from_str`].
-#[derive(Default)]
+/// The `YamlLoader` may load raw YAML documents or add metadata if needed. The type of the `Node`
+/// dictates what data and metadata the loader will add to the `Node`.
+///
+/// Each node must implement [`LoadableYamlNode`]. The methods are required for the loader to
+/// manipulate and populate the `Node`.
 #[allow(clippy::module_name_repetitions)]
-pub struct YamlLoader {
+pub struct YamlLoader<Node>
+where
+    Node: LoadableYamlNode,
+{
     /// The different YAML documents that are loaded.
-    docs: Vec<Yaml>,
+    docs: Vec<Node>,
     // states
     // (current node, anchor_id) tuple
-    doc_stack: Vec<(Yaml, usize)>,
-    key_stack: Vec<Yaml>,
-    anchor_map: BTreeMap<usize, Yaml>,
+    doc_stack: Vec<(Node, usize)>,
+    key_stack: Vec<Node>,
+    anchor_map: BTreeMap<usize, Node>,
 }
 
-impl MarkedEventReceiver for YamlLoader {
-    fn on_event(&mut self, ev: Event, _: Marker) {
-        // println!("EV {:?}", ev);
+// For some reason, rustc wants `Node: Default` if I `#[derive(Default)]`.
+impl<Node> Default for YamlLoader<Node>
+where
+    Node: LoadableYamlNode,
+{
+    fn default() -> Self {
+        Self {
+            docs: vec![],
+            doc_stack: vec![],
+            key_stack: vec![],
+            anchor_map: BTreeMap::new(),
+        }
+    }
+}
+
+impl<Node> MarkedEventReceiver for YamlLoader<Node>
+where
+    Node: LoadableYamlNode,
+{
+    fn on_event(&mut self, ev: Event, marker: Marker) {
         match ev {
             Event::DocumentStart | Event::Nothing | Event::StreamStart | Event::StreamEnd => {
                 // do nothing
@@ -31,21 +55,29 @@ impl MarkedEventReceiver for YamlLoader {
             Event::DocumentEnd => {
                 match self.doc_stack.len() {
                     // empty document
-                    0 => self.docs.push(Yaml::BadValue),
+                    0 => self
+                        .docs
+                        .push(Node::from_bare_yaml(Yaml::BadValue).with_marker(marker)),
                     1 => self.docs.push(self.doc_stack.pop().unwrap().0),
                     _ => unreachable!(),
                 }
             }
             Event::SequenceStart(aid, _) => {
-                self.doc_stack.push((Yaml::Array(Vec::new()), aid));
+                self.doc_stack.push((
+                    Node::from_bare_yaml(Yaml::Array(Vec::new())).with_marker(marker),
+                    aid,
+                ));
             }
             Event::SequenceEnd => {
                 let node = self.doc_stack.pop().unwrap();
                 self.insert_new_node(node);
             }
             Event::MappingStart(aid, _) => {
-                self.doc_stack.push((Yaml::Hash(Hash::new()), aid));
-                self.key_stack.push(Yaml::BadValue);
+                self.doc_stack.push((
+                    Node::from_bare_yaml(Yaml::Hash(Hash::new())).with_marker(marker),
+                    aid,
+                ));
+                self.key_stack.push(Node::from_bare_yaml(Yaml::BadValue));
             }
             Event::MappingEnd => {
                 self.key_stack.pop().unwrap();
@@ -90,18 +122,53 @@ impl MarkedEventReceiver for YamlLoader {
                     // Datatype is not specified, or unrecognized
                     Yaml::from_str(&v)
                 };
-
-                self.insert_new_node((node, aid));
+                self.insert_new_node((Node::from_bare_yaml(node).with_marker(marker), aid));
             }
             Event::Alias(id) => {
                 let n = match self.anchor_map.get(&id) {
                     Some(v) => v.clone(),
-                    None => Yaml::BadValue,
+                    None => Node::from_bare_yaml(Yaml::BadValue),
                 };
-                self.insert_new_node((n, 0));
+                self.insert_new_node((n.with_marker(marker), 0));
             }
         }
-        // println!("DOC {:?}", self.doc_stack);
+    }
+}
+
+impl<Node> YamlLoader<Node>
+where
+    Node: LoadableYamlNode,
+{
+    fn insert_new_node(&mut self, node: (Node, usize)) {
+        // valid anchor id starts from 1
+        if node.1 > 0 {
+            self.anchor_map.insert(node.1, node.0.clone());
+        }
+        if self.doc_stack.is_empty() {
+            self.doc_stack.push(node);
+        } else {
+            let parent = self.doc_stack.last_mut().unwrap();
+            let parent_node = &mut parent.0;
+            if parent_node.is_array() {
+                parent_node.array_mut().push(node.0);
+            } else if parent_node.is_hash() {
+                let cur_key = self.key_stack.last_mut().unwrap();
+                // current node is a key
+                if cur_key.is_badvalue() {
+                    *cur_key = node.0;
+                // current node is a value
+                } else {
+                    let hash = parent_node.hash_mut();
+                    hash.insert(cur_key.take(), node.0);
+                }
+            }
+        }
+    }
+
+    /// Return the document nodes from `self`, consuming it in the process.
+    #[must_use]
+    pub fn into_documents(self) -> Vec<Node> {
+        self.docs
     }
 }
 
@@ -142,76 +209,93 @@ impl std::fmt::Display for LoadError {
     }
 }
 
-impl YamlLoader {
-    fn insert_new_node(&mut self, node: (Yaml, usize)) {
-        // valid anchor id starts from 1
-        if node.1 > 0 {
-            self.anchor_map.insert(node.1, node.0.clone());
-        }
-        if self.doc_stack.is_empty() {
-            self.doc_stack.push(node);
-        } else {
-            let parent = self.doc_stack.last_mut().unwrap();
-            match *parent {
-                (Yaml::Array(ref mut v), _) => v.push(node.0),
-                (Yaml::Hash(ref mut h), _) => {
-                    let cur_key = self.key_stack.last_mut().unwrap();
-                    // current node is a key
-                    if cur_key.is_badvalue() {
-                        *cur_key = node.0;
-                    // current node is a value
-                    } else {
-                        let mut newkey = Yaml::BadValue;
-                        std::mem::swap(&mut newkey, cur_key);
-                        h.insert(newkey, node.0);
-                    }
-                }
-                _ => unreachable!(),
-            }
-        }
-    }
-
-    /// Load the given string as a set of YAML documents.
+/// A trait providing methods used by the [`YamlLoader`].
+///
+/// This trait must be implemented on YAML node types (i.e.: [`Yaml`] and annotated YAML nodes). It
+/// provides the necessary methods for [`YamlLoader`] to load data into the node.
+pub trait LoadableYamlNode: Clone + std::hash::Hash + Eq {
+    /// Create an instance of `Self` from a [`Yaml`].
     ///
-    /// The `source` is interpreted as YAML documents and is parsed. Parsing succeeds if and only
-    /// if all documents are parsed successfully. An error in a latter document prevents the former
-    /// from being returned.
-    /// # Errors
-    /// Returns `ScanError` when loading fails.
-    pub fn load_from_str(source: &str) -> Result<Vec<Yaml>, ScanError> {
-        Self::load_from_iter(source.chars())
-    }
-
-    /// Load the contents of the given iterator as a set of YAML documents.
+    /// Nodes must implement this to be built. The optional metadata that they contain will be
+    /// later provided by the loader and can be default initialized. The [`Yaml`] object passed as
+    /// parameter may be of the [`Array`] or [`Hash`] variants. In this event, the inner container
+    /// will always be empty. There is no need to traverse all elements to convert them from
+    /// [`Yaml`] to `Self`.
     ///
-    /// The `source` is interpreted as YAML documents and is parsed. Parsing succeeds if and only
-    /// if all documents are parsed successfully. An error in a latter document prevents the former
-    /// from being returned.
-    /// # Errors
-    /// Returns `ScanError` when loading fails.
-    pub fn load_from_iter<I: Iterator<Item = char>>(source: I) -> Result<Vec<Yaml>, ScanError> {
-        let mut parser = Parser::new(source);
-        Self::load_from_parser(&mut parser)
-    }
+    /// [`Array`]: `Yaml::Array`
+    /// [`Hash`]: `Yaml::Hash`
+    fn from_bare_yaml(yaml: Yaml) -> Self;
 
-    /// Load the contents from the specified Parser as a set of YAML documents.
+    /// Return whether the YAML node is an array.
+    fn is_array(&self) -> bool;
+
+    /// Return whether the YAML node is a hash.
+    fn is_hash(&self) -> bool;
+
+    /// Return whether the YAML node is `BadValue`.
+    fn is_badvalue(&self) -> bool;
+
+    /// Retrieve the array variant of the YAML node.
     ///
-    /// Parsing succeeds if and only if all documents are parsed successfully.
-    /// An error in a latter document prevents the former from being returned.
-    /// # Errors
-    /// Returns `ScanError` when loading fails.
-    pub fn load_from_parser<I: Iterator<Item = char>>(
-        parser: &mut Parser<I>,
-    ) -> Result<Vec<Yaml>, ScanError> {
-        let mut loader = YamlLoader::default();
-        parser.load(&mut loader, true)?;
-        Ok(loader.docs)
-    }
+    /// # Panics
+    /// This function panics if `self` is not an array.
+    fn array_mut(&mut self) -> &mut Vec<Self>;
 
-    /// Return a reference to the parsed Yaml documents.
+    /// Retrieve the hash variant of the YAML node.
+    ///
+    /// # Panics
+    /// This function panics if `self` is not a hash.
+    fn hash_mut(&mut self) -> &mut LinkedHashMap<Self, Self>;
+
+    /// Take the contained node out of `Self`, leaving a `BadValue` in its place.
     #[must_use]
-    pub fn documents(&self) -> &[Yaml] {
-        &self.docs
+    fn take(&mut self) -> Self;
+
+    /// Provide the marker for the node (builder-style).
+    #[inline]
+    #[must_use]
+    fn with_marker(self, _: Marker) -> Self {
+        self
+    }
+}
+
+impl LoadableYamlNode for Yaml {
+    fn from_bare_yaml(yaml: Yaml) -> Self {
+        yaml
+    }
+
+    fn is_array(&self) -> bool {
+        matches!(self, Yaml::Array(_))
+    }
+
+    fn is_hash(&self) -> bool {
+        matches!(self, Yaml::Hash(_))
+    }
+
+    fn is_badvalue(&self) -> bool {
+        matches!(self, Yaml::BadValue)
+    }
+
+    fn array_mut(&mut self) -> &mut Vec<Self> {
+        if let Yaml::Array(x) = self {
+            x
+        } else {
+            panic!("Called array_mut on a non-array");
+        }
+    }
+
+    fn hash_mut(&mut self) -> &mut LinkedHashMap<Self, Self> {
+        if let Yaml::Hash(x) = self {
+            x
+        } else {
+            panic!("Called hash_mut on a non-hash");
+        }
+    }
+
+    fn take(&mut self) -> Self {
+        let mut taken_out = Yaml::BadValue;
+        std::mem::swap(&mut taken_out, self);
+        taken_out
     }
 }